requirements.txt
asgiref==3.8.1
Django==5.1.4
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
environ==1.0
Markdown==3.7
mysql-connector==2.2.9
pillow==11.1.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
PyJWT==2.10.1
python-decouple==3.8
pytz==2024.1
pywin32==306
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.3
tabulate==0.9.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.2
urllib3==2.2.1
wcwidth==0.2.13
